# api/docs/swagger.yaml
definitions:
...
  routes.responseMessage:
    properties:
      message:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /dashboard/configs/columns:
    patch:
      description: Update the dashboard columns in the configs.json file.
      parameters:
      - description: New number of columns.
        example: 5
        in: query
        name: columns
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.responseMessage'
      summary: Update dashboard columns
...
swagger: "2.0"